Move map layer runtime out of components/ into services/layers
The eight *-layer.ts files under components/map/layers/ contain no templates.
They are singleton runtime classes holding watch handles, rAF registrations and
store references, and their only consumer is
services/layers/markers/marker-layers.service.ts — the same job that
services/layers/features/*.service.ts already does from services/.

Marker layers go to services/layers/markers/; transit-vehicles-layer, which
renders a feature rather than markers, goes to services/layers/features/.
components/map/layers/ now holds only the layer selector UI.
