+++ web/src/services/layers/markers/friend-locations-layer.ts
